[#1049] New dialog for confirming user wants to reload all forms
